+++ Win32ApplicationFramework/lib/source/win32/win32_platform.cpp
@@ -1,7 +1,7 @@
 /*!
 lib\source\win32\win32_platform.cpp
 Created: October 5, 2025
-Updated: November 20, 2025
+Updated: November 22, 2025
 Copyright (c) 2025, Jacob Gosse
 
 Win32 Platform source file.
@@ -12,10 +12,10 @@ Win32 Platform is the main (wWinMain) entry point for the application.
 
 #include <win32/Console/Console.hpp>
 #include <win32/Window/Window.hpp>
-#include <win32/Error/error_macros.hpp>
+#include <win32/Window/window_manager.hpp>
+#include <win32/Error/Error.hpp>
 #include <TestSuite/TestSuite.hpp>
 #include <win32/utils/win32_utils.hpp>
-#include <utils/string_utils.hpp>
 
 using namespace winxframe;
 
@@ -25,25 +25,29 @@ class TestWindow : public Window
 public:
     TestWindow(
         const std::wstring& title = L"",
-        LONG windowWidth = WINDOW_WIDTH,
-        LONG windowHeight = WINDOW_HEIGHT,
-        MessagePumpMode mode = MessagePumpMode::RealTime
+        LONG windowWidth = DEFAULT_SCREEN_WIDTH,
+        LONG windowHeight = DEFAULT_SCREEN_HEIGHT,
+        MessagePumpMode mode = MessagePumpMode::RealTime,
+        int nCmdShow = SW_SHOWDEFAULT
     )
-        : Window(title, windowWidth, windowHeight, mode) {}
+        : Window(title, windowWidth, windowHeight, mode, nCmdShow) {}
 
     TestWindow(
         HINSTANCE hInstance,
         const std::wstring& title = L"",
-        LONG windowWidth = WINDOW_WIDTH,
-        LONG windowHeight = WINDOW_HEIGHT,
-        MessagePumpMode mode = MessagePumpMode::RealTime
+        LONG windowWidth = DEFAULT_SCREEN_WIDTH,
+        LONG windowHeight = DEFAULT_SCREEN_HEIGHT,
+        MessagePumpMode mode = MessagePumpMode::RealTime,
+        int nCmdShow = SW_SHOWDEFAULT
     )
-        : Window(hInstance, title, windowWidth, windowHeight, mode) {}
+        : Window(hInstance, title, windowWidth, windowHeight, mode, nCmdShow) {}
 
     ~TestWindow() override
     {
         std::wcout << L"DESTRUCTOR: ~MyWindow()\n";
     }
+    
+    LRESULT OnCreate() override { return Window::OnCreate(); }
 
     void Update(std::chrono::nanoseconds deltaTime) override
     {
@@ -61,13 +65,14 @@ public:
         win32_utils::RenderFPS(GetMemoryDC(), GetFPS());
         Present();
     }
+
+    LRESULT OnDestroy() override { return Window::OnDestroy(); }
 };
 
 int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance, _In_ LPWSTR lpCmdLine, _In_ int nCmdShow)
 {
     UNREFERENCED_PARAMETER(hPrevInstance);
     UNREFERENCED_PARAMETER(lpCmdLine);
-    UNREFERENCED_PARAMETER(nCmdShow);
 
     // enable memory leak checking
     #if defined(_DEBUG) && defined(_WIN32)
@@ -89,9 +94,9 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance
         // init window(s)
         LONG windowWidth = 800;
         LONG windowHeight = 600;
-        windows.push_back(std::make_unique<TestWindow>(hInstance, L"WINDOW1", windowWidth, windowHeight, MessagePumpMode::RealTime));
-        //windows.push_back(std::make_unique<TestWindow>(hInstance, L"WINDOW2", 600, 400, MessagePumpMode::RealTime));
-        //windows.push_back(std::make_unique<TestWindow>(hInstance, L"WINDOW3", 400, 200, MessagePumpMode::EventDriven));
+        windows.push_back(std::make_unique<TestWindow>(hInstance, L"WINDOW1", windowWidth, windowHeight, MessagePumpMode::RealTime, nCmdShow));
+        //windows.push_back(std::make_unique<TestWindow>(hInstance, L"WINDOW2", 600, 400, MessagePumpMode::RealTime, nCmdShow));
+        //windows.push_back(std::make_unique<TestWindow>(hInstance, L"WINDOW3", 400, 200, MessagePumpMode::EventDriven, nCmdShow));
 
         // run unit tests
         winxframe::TestRegistry::RunAll();
@@ -144,7 +149,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance
     catch (const std::exception& e)
     {
         MessageBoxExA(nullptr, e.what(), "Standard Exception!", MB_OK | MB_ICONERROR, LANG_USER_DEFAULT);
-        std::wcerr << L"Caught Error (std::exception):\n" << winxframe::string_utils::ToWide(e.what()) << L'\n';
+        std::cerr << "Caught Error (std::exception):\n" << e.what() << '\n';
         std::wcout << L"Program exit failure. Press any key to continue..." << std::endl;
         _getch();
         console.reset();
@@ -154,7 +159,7 @@ int APIENTRY wWinMain(_In_ HINSTANCE hInstance, _In_opt_ HINSTANCE hPrevInstance
     catch (const char* msg)
     {
         MessageBoxExA(nullptr, msg, "C-string Exception!", MB_OK | MB_ICONERROR, LANG_USER_DEFAULT);
-        std::wcerr << L"Caught Error (C-string exception):\n" << winxframe::string_utils::ToWide(msg) << L'\n';
+        std::cerr << "Caught Error (C-string exception):\n" << msg << '\n';
         std::wcout << L"Program exit failure. Press any key to continue..." << std::endl;
         _getch();
         console.reset();